test: Add tests for TypeScript definition file
These tests are in-line with the way DefinitelyTyped and other typing
providers test their files, essentially just a small typescript file
with an example in it. If the definition file contains invalid
TypeScript code, or violates some of the constraints of usage, you will
see the test fail.

I've also added it as a component of the Travis CI test process, run
after linting, to ensure that we don't ship a broken definition file.

Obviously this doesn't protect against API changes, I'm still not sure
how to achieve that in an easy way.