refactor: Rewrite TypeScript definition file to ensure it passes tests
The original definition file, while valid TypeScript and valid for
importing as an ambient/global module did not work correctly with the
"node" module resolution mode. I've rewritten it to function correctly
under this mode, removing the need to manually reference it anywhere.

In addition to this, I've explicitly defined return values for all
functions (primarily `void`) to make it possible to run more stringent
checks on the validity of the definition file.