Add TypeScript Definition Tests (#141)
* test: Add tests for TypeScript definition file

These tests are in-line with the way DefinitelyTyped and other typing
providers test their files, essentially just a small typescript file
with an example in it. If the definition file contains invalid
TypeScript code, or violates some of the constraints of usage, you will
see the test fail.

I've also added it as a component of the Travis CI test process, run
after linting, to ensure that we don't ship a broken definition file.

Obviously this doesn't protect against API changes, I'm still not sure
how to achieve that in an easy way.

* refactor: Rewrite TypeScript definition file to ensure it passes tests

The original definition file, while valid TypeScript and valid for
importing as an ambient/global module did not work correctly with the
"node" module resolution mode. I've rewritten it to function correctly
under this mode, removing the need to manually reference it anywhere.

In addition to this, I've explicitly defined return values for all
functions (primarily `void`) to make it possible to run more stringent
checks on the validity of the definition file.