ci: pin docker image to 1.8
Although node-influx supports a subset of the 2.x functionality we
are not currently supporting it. PRs welcome to pass the 2.x tests
but it's unlikely maintainers will be working on this anytime soon.

If you're looking for a commercially supported library I reccomend
taking a look at InfluxData's 2.0 JavaScript client at
https://github.com/influxdata/influxdb-client-js.