feat: add capability to query with POST requests (#717)
* feat: add capability to query with POST requests

This feature allows users to invoke query() and queryRaw() as POST
requests. Closes #676

* fix: apply Content-Type header only for POST queries

Ensures that other operations that were potentially setting a body to
behave as before.

* test: increase timeout for pool unit tests

Pool unit tests were taking longer to finish than the default timeout
allowed them to.

* fix: pass auth credentials as query params

When doing a query with a POST request, Influx expects the username and
password as query parameters and not as part of the POST body.